parser-backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
PyMuPDF
pytesseract
Pillow
psycopg2-binary
python-multipart
